Update CategUpdate to find the Categ with id_server not the id. Create new Function to get The Categ with this id_server update CategWsAdapter to return the list of Categ Update Categ entity get id_server to return int Delete test recup inside the HomeActivity Update HomeFragment to get the list of Categ in DB and manage DB and list with

+++ my_qcm/app/src/main/java/com/tactfactory/my_qcm/entity/Categ.java
@@ -60,7 +60,7 @@ public class Categ {
      * Return id_server of the Categ
      * @return id_server
      */
-    public long getId_server() {
+    public int getId_server() {
         return id_server;
     }
